requirements.txt
aiohttp==3.6.2
async-timeout==3.0.1
attrs==19.3.0
beautifulsoup4==4.9.0
certifi==2020.4.5.1
cffi==1.14.0
chardet==3.0.4
discord.py==1.3.3
fake-useragent==0.1.11
future==0.18.2
Google-Search-API==1.1.14
idna==2.9
idna-ssl==1.1.0
multidict==4.7.5
pycparser==2.20
PyNaCl==1.3.0
python-decouple==3.3
PyYAML==5.3.1
requests==2.23.0
selenium==2.53.6
six==1.14.0
soupsieve==2.0
typing-extensions==3.7.4.2
Unidecode==1.1.1
urllib3==1.25.9
vcrpy==4.0.2
websockets==8.1
wrapt==1.12.1
yarl==1.4.2
youtube-dl==2020.3.24
